+++ CHANGES.md
@@ -8,6 +8,8 @@ Improvements 🙌:
  - The initial sync is now handled by a foreground service
  - Render aliases and canonical alias change in the timeline
  - Fix autocompletion issues and add support for rooms and groups
+ - Introduce developer mode in the settings (#796)
+ - Improve devices list screen
 
 Other changes:
  -